+++ src/config/mod.rs
@@ -1,5 +1,5 @@
 // DKIM Milter – milter for DKIM signing and verification
-// Copyright © 2022–2023 David Bürgin <dbuergin@gluet.ch>
+// Copyright © 2022–2024 David Bürgin <dbuergin@gluet.ch>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software